refactor: PerformanceRepository class -> interface로 리팩토링, MybatisPerformanceRepository 구현
